pyproject.toml
[tool.poetry]
name = "ClassificationPointNet3D"
version = "0.1.0"
description = "Classification ModelNet10 based on 3D vision"
authors = ["Dmitry <alexalex68t@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.27.0"
numpy = "^1.26.2"
torch = "^2.1.1"
torchvision = "^0.16.1"
scipy = "^1.11.4"
plotly = "^5.18.0"
pandas = "^2.1.3"
matplotlib = "^3.8.2"
scikit-learn = "^1.3.2"
kaggle = "^1.5.16"
nbformat = "^5.9.2"
hydra-core = "1.1.0.rc1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
